enhancement(common,trading,test): Stricter behavior with get_bonding_curve_account: raises exception instead of returning None when the bonding curve account cannot be found
