+++ libs/common/solbot_common/utils/utils.py
@@ -1,3 +1,4 @@
+import asyncio
 from decimal import Decimal
 from functools import cache
 
@@ -5,32 +6,34 @@ from jupiter_python_sdk.jupiter import Jupiter
 from loguru import logger
 from solana.rpc.api import Client
 from solana.rpc.async_api import AsyncClient
-from solbot_common.layouts.bonding_curve_account import BondingCurveAccount
-from solbot_common.layouts.global_account import GlobalAccount
-from solbot_common.layouts.mint_account import MintAccount
 from solders.keypair import Keypair  # type: ignore
 from solders.pubkey import Pubkey  # type: ignore
 from solders.signature import Signature  # type: ignore
-from solders.transaction_status import \
-    TransactionConfirmationStatus  # type: ignore
+from solders.transaction_status import TransactionConfirmationStatus  # type: ignore
 from spl.token.instructions import get_associated_token_address
+
+from solbot_common.exceptions import BondingCurveNotFound
+from solbot_common.layouts.bonding_curve_account import BondingCurveAccount
+from solbot_common.layouts.global_account import GlobalAccount
+from solbot_common.layouts.mint_account import MintAccount
 
 
 def get_bonding_curve_pda(mint: Pubkey, program: Pubkey) -> tuple[Pubkey, int]:
     """
     Derives the associated bonding curve Program Derived Address (PDA) for a given mint.
-    
+
     Args:
         mint: The token mint address
         program_id: The program ID for the bonding curve
-        
+
     Returns:
         Tuple of (bonding curve address, bump seed)
     """
 
     return Pubkey.find_program_address([b"bonding-curve", bytes(mint)], program)
-    # return Pubkey.find_program_address([b"creator-vault", bytes(mint)], program)
-def get_bonding_curve_pda_creator_vault(mint: Pubkey, program: Pubkey) :
+
+
+def get_bonding_curve_pda_creator_vault(mint: Pubkey, program: Pubkey):
     """
     Derives the associated bonding curve Program Derived Address (PDA) for a given mint.
 
@@ -45,21 +48,55 @@ def get_bonding_curve_pda_creator_vault(mint: Pubkey, program: Pubkey) :
     # return Pubkey.find_program_address([b"bonding-curve", bytes(mint)], program)
     return Pubkey.find_program_address([b"creator-vault", bytes(mint)], program)
 
+
 async def get_bonding_curve_account(
-    client: AsyncClient, mint: Pubkey, program: Pubkey
-) -> tuple[Pubkey, Pubkey, BondingCurveAccount] | None:
+    client: AsyncClient,
+    mint: Pubkey,
+    program: Pubkey,
+    num_retries: int = 1,
+    retry_delay: float = 0.0,
+) -> tuple[Pubkey, Pubkey, BondingCurveAccount]:
+    """
+    Get the bonding curve account for a given mint
+
+    Args:
+        client: The Solana RPC client
+        mint: The token mint address
+        program: The program ID for the bonding curve
+        num_retries: The number of retries to get the bonding curve account,
+            if the account is not found, it will retry num_retries times
+        retry_delay: The delay between retries in seconds
+
+    Returns:
+        Tuple of (bonding curve PDA, associated token address, bonding curve account)
+    Raises:
+        BondingCurveNotFound: If the bonding curve account is not found after num_retries retries
+    """
+    logger.debug(f"Getting bonding curve account for mint {mint} and program {program}")
     bonding_curve, bump = get_bonding_curve_pda(mint, program)
     associated_bonding_curve = get_associated_token_address(bonding_curve, mint)
 
-    account_info = await client.get_account_info_json_parsed(bonding_curve)
-    if account_info is None:
-        return None
-    value = account_info.value
+    value = None
+    while True:
+        account_info = await client.get_account_info_json_parsed(bonding_curve)
+        if account_info is None:
+            logger.debug("RPC call to get bonding curve account failed.")
+            if (num_retries := num_retries - 1) == 0:
+                break
+            await asyncio.sleep(retry_delay)
+            continue
+        value = account_info.value
+        if value is None:
+            logger.debug("Bonding curve account has no info.")
+            if (num_retries := num_retries - 1) == 0:
+                break
+            await asyncio.sleep(retry_delay)
+            continue
+
     if value is None:
-        return None
+        raise BondingCurveNotFound(f"Bonding curve account not found for mint {mint}")
     bonding_curve_account = BondingCurveAccount(bytes(value.data))
     return bonding_curve, associated_bonding_curve, bonding_curve_account
-
 
 
 async def get_global_account(client: AsyncClient, program: Pubkey) -> GlobalAccount | None:
@@ -105,14 +142,14 @@ def get_jupiter_client() -> Jupiter:
 async def validate_transaction(
     tx_hash: str | Signature, client: AsyncClient | None = None
 ) -> bool | None:
-    """验证交易是否已经上链
+    """Verify that the transaction has been linked
 
     Args:
         tx_hash (str): 交易 hash
         client (AsyncClient): Solana RPC 客户端
 
     Returns:
-        Optional[bool]: None 表示未找到交易或者交易尚未上链，True 表示交易已上链，False 表示交易上链失败
+        Optional[bool]: None † It means that the transaction has been put on the chain, False It means that the transaction failed to be opened
     """
     if client is None:
         client = get_async_client()
